# api/openapi.yaml
openapi: 3.1.0
info:
  title: MailPulse API
  version: 1.0.0
  summary: Watch mailboxes, match messages against filters, and fire events when they match.
  description: |
    MailPulse connects mail accounts over IMAP (or OAuth, per provider), runs each
    incoming message against a watcher's filters, and dispatches the watcher's
    events — notifications, webhooks, HTTP requests — when a message matches.

    ## Response envelope

    Every JSON response is wrapped in the same envelope (`model.WebResponse[T]`):

    ```json
    { "data": <payload>, "paging": { ... }, "errors": "..." }
    ```

    `paging` appears only on paginated list endpoints. `errors` appears only on
    failures, and is the sole field on an error response — the error handler in
    `internal/config/fiber.go` emits `{"errors": "<message>"}` with the matching
    HTTP status. `data` is absent on errors.

    ## Timestamps

    Every timestamp in this API is an integer count of **milliseconds since the
    Unix epoch**, never an RFC 3339 string. The `from` and `to` filters on list
    endpoints take the same unit.

    ## Authentication

    Send the session token from `POST /api/users/_login` as a bearer token:

    ```
    Authorization: Bearer <token>
    ```

    The auth middleware also accepts the bare token with no `Bearer ` prefix.
    Tokens are cached in Redis against their hash; revoking a session or changing
    a user's roles evicts the cached entry immediately.

...

...
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /api/mail-accounts/oauth/{provider}/_authorize:
    parameters:
      - $ref: '#/components/parameters/Provider'
    get:
      tags: [Mail Accounts]
      summary: Begin an OAuth connect flow
      description: |
        Returns the provider's consent URL to redirect the browser to, plus the
        opaque `state` that ties the callback back to this user. Send the user
        to `redirect_url`; the provider then calls `_callback`.

        The state is single use and expires after `REDIS_TTL_OAUTH_STATE`
        seconds, so start the flow when the user is ready to complete it.
      operationId: oauthAuthorize
      responses:
        '200':
          description: Where to send the user, and the state to expect back.
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    $ref: '#/components/schemas/OAuthAuthorizeResponse'
        '400':
          description: |
            The provider is unknown, is not enabled, or does not connect over
            OAuth at all — offer the app-password form instead.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '501':
          description: |
            The provider could connect over OAuth, but this deployment has no
            client id and secret configured for it. The message names the
            app-password fallback.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/mail-accounts/oauth/{provider}/_callback:
    parameters:
      - $ref: '#/components/parameters/Provider'
    get:
      tags: [Mail Accounts]
      summary: Complete an OAuth connect flow
      description: |
        Called by the provider, not by your client. Authenticated by the `state`
        issued at `_authorize` rather than by a session, which is why this route
        needs no bearer token. Exchanges `code` for tokens, encrypts them, and
        creates or updates the mail account.
      operationId: oauthCallback
      security: []
      parameters:
        - name: code
          in: query
          required: true
          description: The authorization code issued by the provider.
          schema: { type: string }
        - name: state
          in: query
          required: true
          description: The value returned by `_authorize`.
          schema: { type: string }
      responses:
        '302':
          description: |
            Always. This endpoint is reached by the provider redirecting a
            browser, so it answers with another redirect rather than a body
            nothing would read.

            On success the browser lands on `<APP_BASE_URL>/mail-accounts?connected=1`.
            On failure it lands on the same page with `?oauth_error=<reason>`,
            where reason is one of `denied` (the user declined at the consent
            screen), `invalid` (the provider sent no code or state), `state`
            (the state was unknown, expired, already spent, or issued for a
            different provider), `unavailable` (the provider is no longer
            configured here), `exchange` (the provider refused the code),
            `identity` (the provider would not say which mailbox this is),
            `mismatch` (a `_reauthorize` came back naming a different mailbox
            than the one it was started for, so nothing was written), or
            `store` (the connection could not be saved).

            The provider's own error text is deliberately never forwarded: it
            is attacker-influenced and would end up reflected into the page.
          headers:
            Location:
              description: Where to send the browser next.
              schema: { type: string, format: uri }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        The `token` from `POST /api/users/_login`. The middleware accepts both
        `Authorization: Bearer <token>` and the bare token.

  # ------------------------------------------------------------- parameters
  parameters:
...
    Provider:
      name: provider
      in: path
      required: true
      description: Provider slug, e.g. `gmail`.
      schema: { type: string, maxLength: 40 }

  # -------------------------------------------------------- shared responses
  responses:
...
    Unauthorized:
      description: The token is missing, malformed, expired, or revoked.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            errors: Unauthorized
...
    NotFound:
      description: |
        No such resource, or it belongs to another user. Ownership failures
        answer `404` rather than `403` so the API does not confirm that an id
        exists to someone who cannot see it.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            errors: Not Found
...
    InternalError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            errors: Internal Server Error

  # ----------------------------------------------------------------- schemas
  schemas:
...
    ErrorResponse:
      type: object
      description: |
        The only body shape on a failure. `data` is absent; `errors` carries the
        message and the HTTP status carries the kind.
      properties:
        errors:
          type: string
          description: Human-readable message. Not a stable machine-readable code.
      required: [errors]
...
    OAuthAuthorizeResponse:
      type: object
      properties:
        redirect_url: { type: string, format: uri, description: Send the user's browser here. }
        state: { type: string, description: Opaque value the provider will return to `_callback`. }
      required: [redirect_url, state]

    # ---- notifiers
...